Tidy up for publishing
- GHC version lowered to 8.10 (base version, default-language)
- Modules moved into Data.* namespace
- Use StrictData and clean up record fields
- Cabal cleanup
- Readme/changelog swapped
