--- src/MyLib.hs
@@ -1,4 +1,0 @@
-module MyLib (someFunc) where
-
-someFunc :: IO ()
-someFunc = putStrLn "someFunc"
