+++ test/Main.hs
@@ -1,4 +1,6 @@
 module Main (main) where
 
+import Data.SparseSet.Boxed qualified as Boxed
+
 main :: IO ()
 main = putStrLn "Test suite not yet implemented."
